Implement switching to named cards
Allows to switch to game ID cards without having to trigger the game ID.

Also allows user to create a card with whatever custom name and be able to switch to it.

Max card name length is 15 characters, names longer than 11 characters are scrolled.

Iteration of named cards is naive, going through entire directory each time you switch the card.
Not great, but should be fine for reasonable amount of cards.

With ~1000 cards in the directory, it was taking ~175ms to switch.
Surely no one is gonna have hundreds of cards, right?

Sort order is whatever ESP8266SdFat does. Maybe in the order of underlying entries?